Fixed a bug in the DefaultMessageProcessor that was causing actor properties to be applied twice when a remote actor was created.  Had to change some methods there to virtual to make it all work out. Also, fixed a warning with the TaskActorTests class between unsigned int and size_t.

+++ trunk/delta3d/tests/dtActors/taskactortests.cpp
@@ -250,7 +250,7 @@ void TaskActorTests::TestTaskSubTasks()
       subTasksGroup->SetValue(*subTasksGroupParam);
       parentProxy->GetAllSubTasks(children);
       CPPUNIT_ASSERT_EQUAL_MESSAGE("Setting the subtasks via the property should populate the list of children.", 
-         size_t(subTasksGroupParam->GetParameterCount()), children.size());
+         subTasksGroupParam->GetParameterCount(), (unsigned int) children.size());
       
       for (unsigned i = 0; i < children.size(); i++)
       {
